clean up env script. fix prod deploy
